changed to float and further optimized ADAM ready for evaluation

+++ C/main.c
@@ -38,10 +38,10 @@ int _main() {
   // matrix_t* a = new_matrix(3,4);
   // matrix_t* b = new_matrix(4,5);
   // for (int i = 0; i < 12; ++i) {
-  //   a->data[i] = (double) i;
+  //   a->data[i] = (float) i;
   // }
   // for (int i = 0; i < 20; ++i) {
-  //   b->data[i] = (double) i;
+  //   b->data[i] = (float) i;
   // }
   // print_matrix(matmul(a, b), 1);
   // run_rl(her_demo);
@@ -75,7 +75,7 @@ static int simple_test() {
   struct timeval start;
   timer_reset(&start);
   run_rl(test);
-  double diff = timer_check(&start);
+  float diff = timer_check(&start);
   printf("Test training took %.1f s\n", diff/1000);
   printf("\n");
   return 0;
